requirements.txt
scikit-learn
flask
streamlit
pandas
numpy
matplotlib
seaborn
ipykernel
PyMySQL
pyyaml
-e .